Replace `tracing-serde` with `tracing-serde-structs`.
