--- tracing-serde-structs/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-name = "tracing-serde-structs"
-version = "0.2.0"
-authors = ["Tokio Contributors <team@tokio.rs>"]
-license = "MIT"
-edition = "2018"
-repository = "https://github.com/tokio-rs/tracing"
-homepage = "https://tokio.rs"
-description = """
-A compatibility layer for serializing trace data with `serde`
-"""
-categories = [
-    "development-tools::debugging",
-    "development-tools::profiling",
-    "asynchronous",
-    "encoding",
-]
-keywords = ["logging", "tracing", "serialization"]
-rust-version = "1.49.0"
-
-[features]
-default = ["std"]
-std = ["serde/std", "tracing-core/std"]
-
-[dependencies]
-serde = { version = "1", default-features = false, features = ["alloc", "derive"] }
-tracing-core = { path = "../tracing-core", version = "0.2", default-features = false }
-heapless = { version = "0.7.10", features = ["serde"] }
-
-[dev-dependencies]
-serde_json = "1"
-
-[badges]
-maintenance = { status = "experimental" }
