+++ tracing-serde/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "tracing-serde"
+name = "tracing-serde-structs"
 version = "0.2.0"
 authors = ["Tokio Contributors <team@tokio.rs>"]
 license = "MIT"
@@ -23,8 +23,9 @@ default = ["std"]
 std = ["serde/std", "tracing-core/std"]
 
 [dependencies]
-serde = { version = "1", default-features = false, features = ["alloc"] }
+serde = { version = "1", default-features = false, features = ["alloc", "derive"] }
 tracing-core = { path = "../tracing-core", version = "0.2", default-features = false }
+heapless = { version = "0.7.10", features = ["serde"] }
 
 [dev-dependencies]
 serde_json = "1"
